+++ build_portable_zip.py
@@ -4,7 +4,7 @@ import os
 import sys
 
 src_dir = "dist/my-dlp"
-out = "dist/my-dlp_v1.3.10_portable.zip"
+out = "dist/my-dlp_v1.3.11_portable.zip"
 
 # Remove any existing file
 if os.path.exists(out):